Messages: move to seedu.address.logic package
The Messages class contains common user-facing messages used by commands
and parsers in the logic layer.

However, Messages is contained within the seedu.address.commons.core
package, which is too general. Given that the Messages class is
user-facing, it would only ever be used by the logic and ui layers, and
thus would fit better inside the seedu.address.logic package.

Furthermore, in the next few commits we will be adding methods to the
Messages class which require access to model classes. This would not be
possible if the Messages class is still in the commons layer, which is
the lowest layer of our architecture and can't access component.

As such, let's move the Messages class to the seedu.address.logic
package.
